+++ packages/infra/esbuild-plugin-gjsify/src/app/gjs.ts
@@ -1,5 +1,6 @@
 import { aliasPlugin } from '@gjsify/esbuild-plugin-alias';
 import { blueprintPlugin } from '@gjsify/esbuild-plugin-blueprint';
+import { cssPlugin } from '@gjsify/esbuild-plugin-css';
 import * as deepkitPlugin from '@gjsify/esbuild-plugin-deepkit';
 import { merge } from "../utils/merge.js";
 import { getAliasesForGjs, globToEntryPoints } from "../utils/index.js";
@@ -94,6 +95,7 @@ export const setupForGjs = async (build: PluginBuild, pluginOptions: PluginOptio
 
     await aliasPlugin(aliases).setup(build);
     await blueprintPlugin().setup(build);
+    await cssPlugin(pluginOptions.css ?? {}).setup(build);
     await deepkitPlugin.deepkitPlugin({reflection: pluginOptions.reflection}).setup(build);
 
     registerToCommonJSPatch(build);